Updating to latest FlutterFlow output.

+++ lib/components/cart_items_widget.dart
@@ -3,7 +3,6 @@ import '/flutter_flow/flutter_flow_count_controller.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
@@ -228,7 +227,8 @@ class _CartItemsWidgetState extends State<CartItemsWidget> {
                                           FFAppState()
                                                   .cart[_model.index]
                                                   .item
-                                                  .price *
+                                                  .price
+                                                  .toDouble() *
                                               FFAppState()
                                                   .cart[_model.index]
                                                   .quantity
